+++ app/database/model/__init__.py
@@ -1,0 +1,1 @@
+from .member import Member
