+++ app/database/model/base.py
@@ -1,0 +1,2 @@
+from sqlalchemy.ext.declarative import declarative_base
+Base = declarative_base()
