+++ app/database/setting.py
@@ -1,0 +1,19 @@
+from sqlalchemy import create_engine
+from sqlalchemy.orm import Session
+from sqlalchemy.ext.declarative import declarative_base
+
+# 接続先DBの設定
+DATABASE = 'sqlite:///circle.sqlite3'
+
+# Engine の作成
+Engine = create_engine(
+  DATABASE,
+  echo=False
+)
+
+# Sessionの作成
+session = Session(
+  autocommit = False,
+  autoflush = True,
+  bind = Engine
+)
