+++ app/eras/src/views/widgets/member_button.py
@@ -1,0 +1,8 @@
+from views.widgets.image_button import ImageButton
+
+class MemberButton(ImageButton):
+    def __init__(self) -> None:
+        super().__init__("member")
+
+    def _on_click(self) -> None:
+        pass
