+++ app/nbot/src/database
@@ -1,0 +1,1 @@
+../../database
