--- projects/eRAS/src/networks/requests/access_request.py
@@ -1,8 +1,0 @@
-from networks.dto_base import DTO
-
-class AccessRequest(DTO):
-    class Place(DTO):
-        campus: str = ""
-        name: str = ""
-    id: str = ""
-    place: Place = None
