--- projects/eRAS/src/networks/requests/student_id.py
@@ -1,4 +1,0 @@
-from networks.dto_base import DTO
-
-class StudentID(DTO):
-    id: int = 0
