--- projects/eRAS/src/views/scene_management/scenes/complete.py
@@ -1,5 +1,0 @@
-from views.scene_management.scenes.announce import Announce
-
-class Complete(Announce):
-    def on_load(self) -> None:
-        return super().on_load()
