--- projects/eRAS/src/views/widgets/setting_button.py
@@ -1,8 +1,0 @@
-from views.widgets.image_button import ImageButton
-
-class SettingButton(ImageButton):
-    def __init__(self) -> None:
-        super().__init__("setting")
-
-    def _on_click(self) -> None:
-        pass
